requirements.txt
robobrowser == 0.5.3
html5lib == 0.9
lxml == 3.4.4
configobj == 5.0.6
Pillow == 2.9.0
tweepy == 3.4.0
